% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check.channels}
\alias{check.channels}
\title{Check if the channels are correctly selected}
\usage{
check.channels(.dirs, .f_par)
}
\description{
The function \code{check_channels} checks that the channel numbers or names
specified in \code{f_par} matches the desired channels in the FCM files.
}
\details{
In order to call \code{preproc_facs_plates} the user must supply a structure
\code{f_par} which contains the names or the numbers of the channels that must
be analyzed. This function is useful to check whether the specified channels
match the ones in the FCM files.
}
