% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary.R
\name{collect_dir}
\alias{collect_dir}
\title{Preprocessing of FCM files.}
\usage{
collect_dir(.dir, .out_dir, .plot, .pdf_dim, .cache_namer, .f_par,
  .filter_preproc_namer, .drop_raw, .use_doParallel, .force)
}
\arguments{
\item{.dir}{Directory to look for the files to preprocess}

\item{.use_doParallel}{If it is set to TRUE then the merging of the results from
all the files in the folder is done parallely on the cores of the local
machine. Please note that this refers only to the final step of the
analysis, when the results of the filtering of the files in the folder must
be merged together; the filtering step is \emph{always} peformed on the
cluster, even when .use_doParallel is set to FALSE. Also note that this
parallelization on the local machine requires the package \emph{doParallel}
and that the user sets up the parallel environment. If the package
doParallel is present and this function is called through
\code{preproc_facs_plates} then the parallel environment is automatically
set to half of the available cores.}
}
\value{
A list containing \enumerate{ \item \emph{stats} the summary
 statistics of all the FCM analized. \item \emph{preproc} the preprocessed
 data, i.e., raw scattering and fluorescence values of the cells which
 survived the scattering filter. If \emph{.drop_raw} is FALSE, then this
 item is NULL.} The function also saves the final result in the file
 specified by \code{.cache_namer} in the folder \code{.out_dir} so that it
 can be restored later if required. Also the function writes the statistiscs
 in a CSV file.
}
\description{
\code{remove_noise} is the main function to preprocess a set of FCM
files contained in one single folder (usually corresponding to a single plate)
after they have been filtered on their scattering profile and their
fluorescence has been fitted with a lognormal distribution through the
function \code{scattering_filter}.
}
\details{
This function is called by \code{remove_noise} after having filtered the
files based on their scattering profile; usually the user shouldn't directly
call this function. It takes all the FCM files in a given folder whose cells
have already been filtered based on their scattering profile and whose results
have been saved in a Rdata structure with the name specified by
\code{.filter_preproc_namer} (this is done with the function
\code{scattering_filter}). Specifically, it merges the filtered files found in
the directory together (possibly in a parallel way) and it removes the
autofluorescence and the shot noise.
}
