% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EM_Luca}
\alias{EM_Luca}
\title{Fit of a mixture of gaussians and one uniform using EM}
\usage{
EM_Luca(y_, mu_, w_, list_sigma_, delta_, verbose_)
}
\arguments{
\item{y_}{The data to be fit with the mixture. It must be a matrix where each
column is a variable and each row is a dimension.}

\item{mu_}{Vector giving the initial values for the means of the gaussians.
Its length determines the number of gaussians to be included in the
mixture.}

\item{w_}{Vector giving the initial wights of the gaussians and the uniform.
It must sum to 1.}

\item{list_sigma_}{List of matrices each giving the starting covariance
matrix for each gaussians.}

\item{delta_}{Dimensions of the uniform distribution. If it's 0 then the
dimensions are the difference between the maximum and the minum of the data
in each direction.}

\item{verbose_}{Should some information be printed?}
}
\value{
A list containing the fitted statistics of the mixture and a matrix
  containing the posterior probability for each point to come from each
  component of the mixture (one column is one component, one row is one data
  point).
}
\description{
This program performs the EM update for a mixture of different gaussians and
one uniform distribution. The width of the uniform if not specified is
computed from the data. The number of gaussians in the mixture is determined
by the length of the parameter \code{mu_}.
}

