% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_autofluo_stats}
\alias{get_autofluo_stats}
\title{Merging different measures supposing that they come form one single gaussian plus some variations}
\usage{
get_autofluo_stats(measures, errors)
}
\arguments{
\item{measures}{The experimental values}

\item{errors}{The error bars on each measures}
}
\value{
A vector containing the estimated mean mu, the standard error of the mean estimate sigma and the width of the variations tau
}
\description{
This function merges the values of different replicates under the following generative model:
It exists a true mean M, nonetheless different replicates have some variation tau around this mean, so that the measured
value on the dataset d is Md=M+tau. This value is measured experimentally with some error bar D, so the final measured value is
M.meas = M + tau + D.
This function estimates M with the error sigma of the estimate, and the width tau of the variations among datasets.
}
