% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{log_stats}
\alias{log_stats}
\title{Mapping of mean and standard deviation between log and linear space.}
\usage{
log_stats(mean, std, to.linear = TRUE)
}
\arguments{
\item{mean}{Mean of the distribution either in linear or log space}

\item{std}{Standard deviation of the distribution either in linear or log space}

\item{to.linear}{Should the mean and standard deviation provided be transformed from linear to log space or viceversa?}
}
\value{
A list containing the transformed mean and standard deviation.
}
\description{
This function maps the mean and standard deviation of a lognormal distribution between log and linear space.
}

