% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_distribution_plots}
\alias{make_distribution_plots}
\title{Plot the fitted fluorescence distributions}
\usage{
make_distribution_plots(.dirs, .data2preproc, .pdf_dim = c(16, 14))
}
\arguments{
\item{.dirs}{Character vector of directories where to look for the processed files created by the function 
\code{analyse_raw}}

\item{.data2preproc}{Function that transforms the directory name of the FCM
files to the directory name where to store preprocessed data.}

\item{.pdf_dim}{Width and height of the pdf to be generated. In inches.}
}
\description{
\code{make_distribution_plots} goes through all the directories and creates a pdf file containing the fluorescence 
distribution of the cells survived to the filtering based on the scattering profile. With each distribution it is also shown 
the fitted mixture of the gaussian and the uniform, from which the mean and variance are extracted. This can be useful to visually
inspect the goodness of the datasets.
}
