% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{Filtering of the scattering profile}
\usage{
preproc(.params)
}
\arguments{
\item{.params}{Dataframe containing 
\enumerate{
\item file: the fsc file to analyze
\item fsc1, fsc2, ssc1, ssc2: the names of the scattering channels
\item gfp: the name of the channel with the GFP intensities
\item out: the name of the output file where to store the statistics
\item scattering_threshold, scattering_frac_cells: the threshold on the posterior or the fraction 
of cells to retain. 
}}
}
\description{
\code{preproc} Is the functiont hat actually performs the fitting of the
Gaussian mixture to the scattering profile and extracts the mean and variance
estimates of the fluorescence of the population.
}
