% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary.R
\name{remove_autofluo}
\alias{remove_autofluo}
\title{Remove autofluorescence}
\usage{
remove_autofluo(.stats, .empty_wells, .plot = FALSE)
}
\value{
The function returns the dataframe .stats with the additional fields fl_mean_lin_autofluo_rm and fl_mean_lin_autofluo_rm.err containing the fluorescence and its 
error in linear space with the autofluorescence removed. It also adds the fields autofluo_mean_lin and autofluo_mean_lin.err which contain the estimated autofluorescence and its error in linear space.
}
\description{
\code{remove_autofluo} removes the autofluorescence from the FCS measurements. 
It does that by considering the fluorescene of empty plasmids. If more than one empty plasmid is 
provided, the values are first filtered to remove outliers (every point which is more than 
q3+0.75IQR or less than q1-0.75IQR, where q1 and q3 are the first and third quartiles and IQRis the interquartile range). The survival points are collapsed together to 
get a single estimate of the autofluorescence and its error.
If only one empty well is given, the removal of outliers is skipped.
}
