% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary.R
\name{remove_noise}
\alias{remove_noise}
\title{Preprocessing of FCM files.}
\usage{
remove_noise(.dir, .out_dir, .f_par, .plot, .verbose, .pdf_dim,
  .cache_namer, .filter_preproc_namer, .drop_raw, .use_doParallel, .force)
}
\arguments{
\item{.dir}{Directory to look for the files to preprocess}

\item{.f_par}{List containing the parameters used to preprocess the data. It
must contain the elements \enumerate{ \item \emph{channels}: the names of
the channels containing the signals to be analysed, in the order: forward
scattering first signal, forward scattering second signal, side scattering
first signal, side scattering second signal, fluorescence. \item
\emph{scattering_threshold}: a cell is considered viable if its probability
to come from the main scattering distribution is bigger or equal to this
threshold. \item \emph{scattering_frac_cells}: retain only the fraction
\emph{scattering_frac_cells} of cells with the highest poterior distribution; this is
a way to specify a specific number of cells to retain from the filtering. If
\emph{scattering_threshold} is specified, \emph{scattering_frac_cells} will
be ignored. \item \emph{file_pattern}: the pattern fo the FCM files to be
analysed. Useful if the directories contain files that must not be analysed.
\item \emph{delta_shot_noise}: the amplitude of the shot noise. 
\item \emph{delta_shot_noise_err}: the error associated to the shot noise. It can be zero\item
\emph{autofluo}: Tibble containing for every dir information about the autofluorescence to be subtracted. 
These are, in order: mean, error of the mean, standard deviation, error of the standard deviation. 
These can be computed with the function get_autofluo_stats()}}

\item{.plot}{Should some plots summarizing the filtering be produced and
stored to a pdf? Default to TRUE
@param .force_noise_rm Forces the removal of the autofluorescence and of the shot noise}

\item{.verbose}{Should some information about the several stages the function
goes through be displayed? Default to FALSE}

\item{.pdf_dim}{Dimensions of the pdf file where to print the plots: c(width,
height, rows, columns)}

\item{.cache_namer}{Function that takes the directory name where to store the
preprocessed files and returns the name and path of a Rdata file where to
store the preprocessed data of all the files in the directory. By default,
if \emph{path} is the directory of the preprocessed data, the file is stored
in \emph{path/path_preproc_luca.Rdata}.}

\item{.filter_preproc_namer}{Function that takes the directory name where to
store the preprocessed files and returns the name and path of a Rdata file
where to store the preprocessed data of the single files. By default, if
\emph{path} is the directory of the preprocessed data and the file is called
abc.fcs, then the preprocessed data for that file is stored in
\emph{path/abc.Rdata}.}

\item{.drop_raw}{Should the raw scattering and fluorescence values of the
cells which survived the scattering filter (i.e. viable cells) be dropped
and only the summary statistics be returned? Dropping them can save space
and make the algorithm faster. Default to TRUE.}

\item{.use_doParallel}{If it is set to TRUE then the merging of the results from
all the files in the folder is done parallely on the cores of the local
machine. Please note that this refers only to the final step of the
analysis, when the results of the filtering of the files in the folder must
be merged together; the filtering step is \emph{always} peformed on the
cluster, even when .use_doParallel is set to FALSE. Also note that this
parallelization on the local machine requires the package \emph{doParallel}
and that the user sets up the parallel environment. If the package
doParallel is present and this function is called through
\code{preproc_facs_plates} then the parallel environment is automatically
set to half of the available cores.}
}
\value{
A list containing \enumerate{ \item \emph{stats} the summary
 statistics of all the FCM analized. \item \emph{preproc} the preprocessed
 data, i.e., raw scattering and fluorescence values of the cells which
 survived the scattering filter. If \emph{.drop_raw} is FALSE, then this
 item is NULL.} The function also saves the final result in the file
 specified by \code{.cache_namer} in the folder \code{.out_dir} so that it
 can be restored later if required. Also the function writes the statistiscs
 in a CSV file.
}
\description{
\code{remove_noise} is the main function to preprocess a set of FCM
files contained in one single folder (usually corresponding to a single plate)
after they have been filtered on their scattering profile and their
fluorescence has been fitted with a lognormal distribution through the
function \code{scattering_filter}.
}
\details{
This function is called by \code{remove_noise} after having filtered the
files based on their scattering profile; usually the user shouldn't directly
call this function. It takes all the FCM files in a given folder whose cells
have already been filtered based on their scattering profile and whose results
have been saved in a Rdata structure with the name specified by
\code{.filter_preproc_namer} (this is done with the function
\code{scattering_filter}). Specifically, it merges the filtered files found in
the directory together (possibly in a parallel way) and it removes the
autofluorescence and the shot noise.
}
