% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary.R
\name{scattering_filter}
\alias{scattering_filter}
\title{Filtering of FCM files based on their scattering profile.}
\usage{
scattering_filter(.dirs, .data2preproc, .f_par, .filter_preproc_namer,
  .preproc_script, .jobs_cmd_name, .force)
}
\value{
The function doesn't return any value, but it saves the statistics of
 the filtered data in Rdata files in the folder \code{.data2preproc} with the
 name given by \code{.filter_preproc_name}.
}
\description{
\code{scattering_filter} filters all the FCM files in a set of folders based
on their scattering profile and it fits the survived cells with a mixture of
lognormal and uniform distribution. It stores the final result for each FCM
file in the directory specified by \code{.data2preproc}, such that they can be
loaded for subsequent analysis (e.g. by \code{remove_noise}).
}
\details{
This function loops over all the dirs provided and check for FCM files that
have not yet been filtered. The filtered information are stored in a path
given by the function .data2preproc and with a filename given by
.filter_preproc_namer. More specificaly, what the function does is to check
for every FCM files if the file storing the filtered info is present. If it is
not present, or if .force is TRUE, it adds the command to filter the file in
the file CMD in the working directory. At the end, it creates a shell job file
to execute an array job which is sent to the cluster. IMPORTANT: dirnames must
be given as absolute paths or as relative the current working directory.
}
