% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{analyse_raw}
\alias{analyse_raw}
\title{Preprocessing of FCM files.}
\usage{
analyse_raw(.dirs, .data2preproc, .filter_preproc_namer = (function(.d)
  sub("fcs$", "RData", .d)), .f_par, .force = FALSE,
  .preproc_func = NULL)
}
\arguments{
\item{.dirs}{Character vector of directories containing the files to be processed.}

\item{.data2preproc}{Function that transforms the directory names of the FCM
files to the directory names where to store the processed data. For example, if the files are in a directory
called \emph{data/condition0/}, you want to store the processed data in \emph{data_processed/condition0/}. In 
order to do that, you can set \code{data2preproc <- function(.d) sub('data', 'data_proc', .d)}}

\item{.filter_preproc_namer}{Function that takes the directory name where to
store the processed files and returns the name and path of a Rdata file
where to store the processed data of the single FCM files. By default, if
\emph{path} is the directory of the processed data and the file is called
abc.fcs, then the processed data for that file is stored in
\emph{path/abc.Rdata}.}

\item{.f_par}{List containing the parameters used to process the FCM files. It
must contain the elements \enumerate{ \item \emph{channels}: the names of
the channels containing the signals to be analysed, in the order: forward
scattering first signal, forward scattering second signal, side scattering
first signal, side scattering second signal, fluorescence. Usually the first 
signal is the height and the second one is the width. \item
\emph{scattering_threshold}: a cell is considered viable if its probability
to come from the main scattering distribution is bigger or equal to this
threshold. \item \emph{scattering_frac_cells}: retain only the fraction
\emph{scattering_frac_cells} of cells with the highest poterior distribution; this is
a way to specify a specific number of cells to retain from the filtering. If
\emph{scattering_threshold} is specified, \emph{scattering_frac_cells} will
be ignored. \item \emph{file_pattern}: the pattern of the FCM files to be
analysed. Useful if the directories contain files that must not be analysed.
\item \emph{delta_shot_noise}: the amplitude of the shot noise. 
\item \emph{delta_shot_noise_err}: the error associated to the shot noise. It can be zero.}}

\item{.force}{Forces the removal of the autofluorescence and of the shot noise, even if the FCM have already been 
processed.}

\item{.preproc_func}{Name of the function that performs the filtering of the
scattering signals and the lognormal fit of the fluorescence. If NULL, the 
default function is used.}
}
\value{
A dataframe containing the summary
 statistics of all the FCM analized \enumerate{\item \emph{n}: the number of cells after the scattering filtering.
 \item \emph{fl_mean, fl_var, fl_mean.err, fl_var.err, fl_mean_lin, fl_mean_lin.err, fl_var_lin, fl_var_lin.err}: 
 estimated means and variances of the populations of cells in log and real space, with their error.
 \item \emph{scattering_threshold, scattering_frac_cells}: parameters used to perform the scattering filtering. They are
 the ones passed to the function through the \emph{.f_par} structure}.
}
\description{
\code{analyse_raw} reads all FCM files in a group of directories, filters them based on the scattering signals, and 
returns the mean and variance of the populations of cells.
}
\details{
This functions is the \emph{main function} that should be called to process
FCM data. It takes all the FCM files in a set of folders and applies the
filtering in the scattering space to remove non viable cells. Then it fits a
mixture of a lognormal and a uniform distribution to extract the mean and
variance from the fluorescence of the remaining population. Notice that to speed up the 
processing, the function \emph{plan} from the package \emph{future} can be called
just before this function.
}
