% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check.channels}
\alias{check.channels}
\title{Check if the channels are correctly selected}
\usage{
check.channels(.dirs, .f_par)
}
\arguments{
\item{.dirs}{List of directories containing the files to be preprocessed.}

\item{.f_par}{List containing the parameters used to preprocess the data. It
must contain the elements \enumerate{ \item \emph{channels}: the names of
the channels containing the signals to be analysed, in the order: forward
scattering first signal, forward scattering second signal, side scattering
first signal, side scattering second signal, fluorescence. \item
\emph{scattering_threshold}: a cell is considered viable if its probability
to come from the main scattering distribution is bigger or equal to this
threshold. \item \emph{scattering_num_cells}: retain only the
\emph{scattering_num_cells} with the highest poterior distribution; this is
a way to specify a specific number of cells to retain from the filtering. If
\emph{scattering_threshold} is specified, \emph{scattering_num_cells} will
be ignored. \item \emph{file_pattern}: the pattern fo the FCM files to be
analysed. Useful if the directories contain files that must not be analysed.
\item \emph{delta_shot_noise}: the amplitude of the shot noise. \item
\emph{autofluo_wells}: Tibble containing for every dir the names of the
empty plasmids wells }}
}
\description{
The function \code{check_channels} checks that the channel numbers or names
specified in \code{f_par} matches the desired channels in the FCM files.
}
\details{
In order to call \code{preproc_facs_plates} the user must supply a structure
\code{f_par} which contains the names or the numbers of the channels that must
be analyzed. This function is useful to check whether the specified channels
match the ones in the FCM files.
}

