% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remove_noise}
\alias{remove_noise}
\title{Remove the autofluorescence and the shot noise}
\usage{
remove_noise(.pls, .f_par)
}
\arguments{
\item{.pls}{The results from \code{analyse_raw} with added information about the autofluorescence. The autofluorescence
information must be stored in the columns: 
\enumerate{\item \emph{autofluo_mean, autofluo_mean.err, autofluo_var, autofluo_var.err}: the mean and variance of the 
autofluorescence and their errors}
See the example file to see how to compute this information}
}
\value{
The same dataframe as the input \emph{.pls}, but with the extra columns
\enumerate{\item \emph{fl_mean_lin_autofluo_rm, fl_var_lin_autofluo_rm}: The mean fluorescence of the population after 
autofluorescence removal. \item \emph{fl_var_lin_noise_rm}: The autofluorescence of the population after removal of 
autofluorescence and shot noise.}
Notice that the mean is not affected by the shot noise removal, therefore we only return the mean corrected by the 
autofluorescence.
}
\description{
\code{remove_noise} is responsible to remove the autofluorescence and the shot noise from the processed data returned by 
\code{analyse_raw}.
}
